+++ website/src/proxy.ts
@@ -11,7 +11,7 @@ const CLOUDFLARE_IP_COUNTRY_HEADER = 'cf-ipcountry';
 export const config = {
 	matcher: [
 		// Skip internal paths (_next)
-		'/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js|portal|partner-space|v1/api-docs|openapi.json|sitemap.xml|new-website/sitemap.xml).*)',
+		'/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js|portal|partner-space|v1/api-docs|openapi.json|sitemap.xml|new-website/sitemap.xml|new-website/robots.txt|new-website/llms.txt).*)',
 	],
 };
